+++ test/suite/Stub/EmptyValueFactoryTest.php
@@ -20,8 +20,20 @@ class EmptyValueFactoryTest extends PHPUnit_Framework_TestCase
 {
     protected function setUp()
     {
+        if (!class_exists('ReflectionType')) {
+            $this->markTestSkipped('Requires reflection types.');
+        }
+
         $this->subject = new EmptyValueFactory();
+        $this->subject->setStubVerifierFactory(StubVerifierFactory::instance());
         $this->subject->setMockBuilderFactory(MockBuilderFactory::instance());
+    }
+
+    private function createType($type)
+    {
+        $reflector = new ReflectionFunction(eval("return function (): $type {};"));
+
+        return $reflector->getReturnType();
     }
 
     public function fromTypeData()
@@ -40,99 +52,221 @@ class EmptyValueFactoryTest extends PHPUnit_Framework_TestCase
      */
     public function testFromType($type, $expected)
     {
-        if (!class_exists('ReflectionType')) {
-            $this->markTestSkipped('Requires reflection types.');
-        }
-
-        $reflector = new ReflectionFunction(eval("return function () : $type {};"));
-        $type = $reflector->getReturnType();
-
-        $this->assertSame($expected, $this->subject->fromType($type));
-    }
-
-    public function testFromTypeWithCallable()
-    {
-        if (!class_exists('ReflectionType')) {
-            $this->markTestSkipped('Requires reflection types.');
-        }
-
-        $reflector = new ReflectionFunction(eval('return function () : callable {};'));
-        $type = $reflector->getReturnType();
-
-        $this->assertInstanceOf('Closure', $this->subject->fromType($type));
+        $this->assertSame($expected, $this->subject->fromType($this->createType($type)));
     }
 
     public function testFromTypeWithStdClass()
     {
-        if (!class_exists('ReflectionType')) {
-            $this->markTestSkipped('Requires reflection types.');
-        }
+        $actual = $this->subject->fromType($this->createType('stdClass'));
 
-        $reflector = new ReflectionFunction(eval('return function () : stdClass {};'));
-        $type = $reflector->getReturnType();
-
-        $this->assertInstanceOf('stdClass', $this->subject->fromType($type));
+        $this->assertSame(array(), (array) $actual);
+        $this->assertSame('{}', json_encode($actual));
     }
 
-    public function testFromTypeWithTraversable()
+    public function testFromTypeWithCallable()
     {
-        if (!class_exists('ReflectionType')) {
-            $this->markTestSkipped('Requires reflection types.');
-        }
+        $actual = $this->subject->fromType($this->createType('callable'));
 
-        $reflector = new ReflectionFunction(eval('return function () : Traversable {};'));
-        $type = $reflector->getReturnType();
-
-        $this->assertInstanceOf('EmptyIterator', $this->subject->fromType($type));
+        $this->assertInstanceOf('Eloquent\Phony\Stub\StubVerifier', $actual);
+        $this->assertNull($actual());
     }
 
-    public function testFromTypeWithIterator()
+    public function testFromTypeWithClosure()
     {
-        if (!class_exists('ReflectionType')) {
-            $this->markTestSkipped('Requires reflection types.');
+        $actual = $this->subject->fromType($this->createType('Closure'));
+
+        $this->assertInstanceOf('Closure', $actual);
+        $this->assertNull($actual());
+    }
+
+    public function fromTypeWithIteratorTypeData()
+    {
+        $types = array(
+            'AppendIterator',
+            'ArrayIterator',
+            'CachingIterator',
+            'CallbackFilterIterator',
+            'DirectoryIterator',
+            'DirectoryIterator',
+            'EmptyIterator',
+            'FilesystemIterator',
+            'FilesystemIterator',
+            'FilterIterator',
+            'GlobIterator',
+            'GlobIterator',
+            'InfiniteIterator',
+            'Iterator',
+            'IteratorIterator',
+            'LimitIterator',
+            'MultipleIterator',
+            'NoRewindIterator',
+            'OuterIterator',
+            'ParentIterator',
+            'RecursiveArrayIterator',
+            'RecursiveCachingIterator',
+            'RecursiveCallbackFilterIterator',
+            'RecursiveDirectoryIterator',
+            'RecursiveDirectoryIterator',
+            'RecursiveFilterIterator',
+            'RecursiveIterator',
+            'RecursiveIteratorIterator',
+            'RecursiveRegexIterator',
+            'RecursiveTreeIterator',
+            'RegexIterator',
+            'SeekableIterator',
+            'Traversable',
+        );
+        $data = array();
+
+        foreach ($types as $type) {
+            $data[$type] = array($type);
         }
 
-        $reflector = new ReflectionFunction(eval('return function () : Iterator {};'));
-        $type = $reflector->getReturnType();
+        return $data;
+    }
 
-        $this->assertInstanceOf('EmptyIterator', $this->subject->fromType($type));
+    /**
+     * @dataProvider fromTypeWithIteratorTypeData
+     */
+    public function testFromTypeWithIteratorType($type)
+    {
+        $actual = $this->subject->fromType($this->createType($type));
+
+        $this->assertInstanceOf($type, $actual);
+        $this->assertInstanceOf('Eloquent\Phony\Mock\Mock', $actual);
+        $this->assertSame(array(), iterator_to_array($actual));
     }
 
     public function testFromTypeWithGenerator()
     {
-        if (!class_exists('ReflectionType')) {
-            $this->markTestSkipped('Requires reflection types.');
-        }
         if (!class_exists('Generator')) {
             $this->markTestSkipped('Requires generators.');
         }
 
-        $reflector = new ReflectionFunction(eval('return function () : Generator {};'));
-        $type = $reflector->getReturnType();
-        $actual = $this->subject->fromType($type);
+        $actual = $this->subject->fromType($this->createType('Generator'));
 
         $this->assertInstanceOf('Generator', $actual);
         $this->assertSame(array(), iterator_to_array($actual));
     }
 
-    public function testFromTypeWithObject()
+    public function fromTypeWithThrowableTypeData()
     {
-        if (!class_exists('ReflectionType')) {
-            $this->markTestSkipped('Requires reflection types.');
+        $types = array(
+            'ArithmeticError',
+            'AssertionError',
+            'BadFunctionCallException',
+            'BadMethodCallException',
+            'DivisionByZeroError',
+            'DomainException',
+            'Error',
+            'ErrorException',
+            'Exception',
+            'InvalidArgumentException',
+            'LengthException',
+            'LogicException',
+            'OutOfBoundsException',
+            'OutOfRangeException',
+            'OverflowException',
+            'ParseError',
+            'PharException',
+            'PDOException',
+            'RangeException',
+            'ReflectionException',
+            'RuntimeException',
+            'Throwable',
+            'TypeError',
+            'UnderflowException',
+            'UnexpectedValueException',
+        );
+        $data = array();
+
+        foreach ($types as $type) {
+            $data[$type] = array($type);
         }
 
-        $reflector = new ReflectionFunction(eval('return function () : Eloquent\Phony\Test\TestClassA {};'));
-        $type = $reflector->getReturnType();
+        return $data;
+    }
 
-        $this->assertInstanceOf('Eloquent\Phony\Test\TestClassA', $this->subject->fromType($type));
+    /**
+     * @dataProvider fromTypeWithThrowableTypeData
+     */
+    public function testFromTypeWithThrowableType($type)
+    {
+        $actual = $this->subject->fromType($this->createType($type));
+
+        $this->assertInstanceOf($type, $actual);
+        $this->assertInstanceOf('Eloquent\Phony\Mock\Mock', $actual);
+        $this->assertSame('', $actual->getMessage());
+        $this->assertSame(0, $actual->getCode());
+        $this->assertNull($actual->getPrevious());
+    }
+
+    public function fromTypeWithCollectionTypeData()
+    {
+        $types = array(
+            'SplDoublyLinkedList',
+            'SplFixedArray',
+            'SplHeap',
+            'SplMaxHeap',
+            'SplMinHeap',
+            'SplObjectStorage',
+            'SplPriorityQueue',
+            'SplQueue',
+            'SplStack',
+        );
+        $data = array();
+
+        foreach ($types as $type) {
+            $data[$type] = array($type);
+        }
+
+        return $data;
+    }
+
+    /**
+     * @dataProvider fromTypeWithCollectionTypeData
+     */
+    public function testFromTypeWithCollectionType($type)
+    {
+        $actual = $this->subject->fromType($this->createType($type));
+
+        $this->assertInstanceOf($type, $actual);
+        $this->assertInstanceOf('Eloquent\Phony\Mock\Mock', $actual);
+        $this->assertSame(array(), iterator_to_array($actual));
+        $this->assertSame(0, count($actual));
+    }
+
+    public function testFromTypeWithArrayAccess()
+    {
+        $type = 'ArrayAccess';
+        $actual = $this->subject->fromType($this->createType($type));
+
+        $this->assertInstanceOf($type, $actual);
+        $this->assertInstanceOf('Eloquent\Phony\Mock\Mock', $actual);
+        $this->assertFalse(isset($actual[0]));
+    }
+
+    public function testFromTypeWithCountable()
+    {
+        $type = 'Countable';
+        $actual = $this->subject->fromType($this->createType($type));
+
+        $this->assertInstanceOf($type, $actual);
+        $this->assertInstanceOf('Eloquent\Phony\Mock\Mock', $actual);
+        $this->assertSame(0, count($actual));
+    }
+
+    public function testFromTypeWithObject()
+    {
+        $type = 'Eloquent\Phony\Test\TestClassA';
+        $actual = $this->subject->fromType($this->createType($type));
+
+        $this->assertInstanceOf($type, $actual);
+        $this->assertInstanceOf('Eloquent\Phony\Mock\Mock', $actual);
     }
 
     public function testFromTypeWithNullableType()
     {
-        if (!class_exists('ReflectionType')) {
-            $this->markTestSkipped('Requires reflection types.');
-        }
-
         $reflector = new ReflectionFunction(eval('return function (int $i = null) {};'));
         $parameters = $reflector->getParameters();
         $type = $parameters[0]->getType();